\name{xMSannotator-package}
\alias{xMSannotator-package}
\alias{xMSannotator}
\docType{package}
\title{
\packageTitle{xMSannotator}
}
\description{
\packageDescription{xMSannotator}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{xMSannotator}

Maintainer: \packageMaintainer{xMSannotator}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = xMSannotator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{xMSannotator}}
%\section{Documentation Index}{\packageIndices{xMSannotator}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
